refactor(dispatcher): extract DispatchRequest logic to dispatcher_request.go
- Moved DispatchRequest and related option types to dispatcher_request.go
- Created helper functions for options creation, token validation, and payload reading
- Simplified DispatchRequest method by reusing helper functions
- Removed unused imports and cleaned up dispatcher.go file
